Refactoring: Rename 'CharacterDirection' enum into 'Direction', change script location and namespace
